iter code reorganization
too large files are broken into pieces.
