+++ src/iter/mod.rs
@@ -1,12 +1,9 @@
 mod dfs;
-mod kind_core;
-mod kind_over;
+mod iter_kinds;
 
 pub use dfs::Dfs;
-pub use kind_core::{
-    DataFromNode, IterKindCore, NodeDepthSiblingVal, NodeDepthVal, NodeFromNode, NodeVal,
-};
-pub use kind_over::{
-    IterOver, OverData, OverDepthData, OverDepthNode, OverDepthSiblingData, OverDepthSiblingNode,
-    OverNode,
+pub use iter_kinds::{
+    DataFromNode, IterKindCore, IterOver, NodeDepthSiblingVal, NodeDepthVal, NodeFromNode, NodeVal,
+    OverData, OverDepthData, OverDepthNode, OverDepthSiblingData, OverDepthSiblingNode, OverNode,
+    StackElement, ValueFromNode,
 };
